+++ packages/pigeon/platform_tests/alternate_language_test_plugin/ios/Classes/AlternateLanguageTestPlugin.m
@@ -45,6 +45,11 @@
   *error = [FlutterError errorWithCode:@"An error" message:nil details:nil];
 }
 
+- (nullable id)throwFlutterErrorWithError:(FlutterError *_Nullable *_Nonnull)error {
+  *error = [FlutterError errorWithCode:@"code" message:@"message" details:@"details"];
+  return nil;
+}
+
 - (nullable NSNumber *)echoInt:(NSNumber *)anInt error:(FlutterError *_Nullable *_Nonnull)error {
   return anInt;
 }
@@ -159,6 +164,11 @@
 
 - (void)throwAsyncErrorFromVoidWithCompletion:(void (^)(FlutterError *_Nullable))completion {
   completion([FlutterError errorWithCode:@"An error" message:nil details:nil]);
+}
+
+- (void)throwAsyncFlutterErrorWithCompletion:(void (^)(id _Nullable,
+                                                       FlutterError *_Nullable))completion {
+  completion(nil, [FlutterError errorWithCode:@"code" message:@"message" details:@"details"]);
 }
 
 - (void)echoAsyncAllTypes:(AllTypes *)everything